stevenflaskblog/requirements.txt
appdirs==1.4.4
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
click==8.0.1
colorama==0.4.4
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.3
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.0
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==8.3.0
pipenv==2021.5.29
pycparser==2.20
pylint==2.6.0
six==1.15.0
SQLAlchemy==1.4.20
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.10.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
